requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
sqlalchemy
passlib[bcrypt]
google-generativeai
